requirements.txt
# sudo apt-get install mysql-server libmysqlclient-dev libpq-dev python-dev python3-dev build-essential 
# dvc init - to track dataset 
# dvc remote add -d storage gdrive://1zOuuO3tVDjAwvJV9Bowk5-hdNFiL93Y1 - set remote storage gdrive
# dvc add /home/asdf/prj/MLFlow/data/LargeMovieReviewDataset.csv - add dataset to the dvc tracking
# dvc push - create link with a dvc remote sta folder with data in gdrive girectory
# rm -f /home/asdf/prj/MLFlow/data/LargeMovieReviewDataset.csv
# rm -rf .dvc/cache
# dvc pull - to pull dataset from gdrive 

wheel
sklearn-pandas
scripts
unidecode
word2number
contractions
pyspellchecker==0.5.6
pattern
beautifulsoup4
contractions
nltk
spacy==2.3.5
unidecode
word2number
matplotlib
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
hyperopt
seaborn
mlflow
gevent
dvc[gdrive]